ENH: allow per-operation connection for thread-safety.
The main use case of this library is integration with
cachecontrol. When a requests.Session is used in multiple threads, the
underlying adapter created by cachecontrol will also be used from
multiple threads, implicitely re-using the same sqlite connection. This
does not work.

Instead, we add a mode to SqliteCache to use a new connection for each
get/set/delete.